requirements.txt
streamlit==1.29.0
openai==1.6.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pypdf2==3.0.1
python-docx==0.8.11
tiktoken==0.5.2
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
plotly==5.17.0